+++ light/light.h
@@ -23,7 +23,7 @@ class Light {
     virtual Vector3d get_contribution(Reflexivity &reflex, Vector3d &l,
                                        Vector3d &n, Vector3d &v,
                                        Vector3d &r) = 0;
-    virtual double get_distance_from_p(Vector3d p_i) = 0;
+    virtual long double get_distance_from_p(Vector3d p_i) = 0;
     virtual void change_some_proprety(gMatrix &wc) = 0;
     virtual std::string toStr(gMatrix &cw) = 0;
     virtual void operator*(gMatrix m) = 0;
