+++ shapes/plane/plane.cpp
@@ -3,12 +3,12 @@
 Plane::Plane(Reflexivity reflexivity, Vector3d p_pi, Vector3d n)
     : Shape(reflexivity, p_pi), p_pi_(p_pi), n_(n){};
 
-double Plane::intersect(Vector3d &p_0, Vector3d &dr) {
-    double t;
+long double Plane::intersect(Vector3d &p_0, Vector3d &dr) {
+    long double t;
 
     Vector3d w = p_0 - p_pi_;
 
-    double den = dr.dot(n_);
+    long double den = dr.dot(n_);
 
     if (den == 0) {
         return INFINITY;
